refactor(python): move the ten inline test modules to sibling files (#1918)
Third crate of the per-crate migration campaign. Pure move via the
#[path] sibling pattern across all ten baselined files; the sorted
test-name list is byte-identical before and after (97 lib tests), run
through the crate's documented PYO3_PYTHON/PYTHONPATH recipe. The
frozen baseline shrinks by the ten velesdb-python lines (1373 lines
paid); 167 files remain.
