+++ crates/velesdb-python/src/graph_store.rs
@@ -434,14 +434,5 @@ fn dfs_collect(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_streaming_config_defaults() {
-        let config = StreamingConfig::new(3, 10000, None);
-        assert_eq!(config.max_depth, 3);
-        assert_eq!(config.max_visited, 10000);
-        assert!(config.relationship_types.is_none());
-    }
-}
+#[path = "graph_store_tests.rs"]
+mod tests;
